Move out methods from anoncreds.rs
