+++ aries_vcx/src/libindy/proofs/verifier/mod.rs
@@ -1,2 +1,3 @@
 pub mod verifier;
 mod verifier_internal;
+mod verifier_libindy;
