+++ libvcx/src/api_lib/api_handle/revocation_registry.rs
@@ -1,7 +1,7 @@
 use aries_vcx::error::prelude::*;
 use aries_vcx::libindy::primitives::revocation_registry::RevocationRegistry;
 use aries_vcx::libindy::anoncreds;
-use aries_vcx::libindy::anoncreds::RevocationRegistryDefinition;
+use aries_vcx::libindy::primitives::revocation_registry::RevocationRegistryDefinition;
 use crate::api_lib::global::pool::get_main_pool_handle;
 
 use crate::api_lib::api_handle::object_cache::ObjectCache;
